feat: add local resolver

+++ src/main/java/com/ia03/service/impl/UserServiceImpl.java
@@ -93,7 +93,12 @@ public class UserServiceImpl implements UserService {
   }
 
   @Override
+  @SneakyThrows
   public void deleteUser(Long id) {
+    if (!userRepository.existsById(id)) {
+      throw new EntityNotFoundException();
+    }
+
     userRepository.deleteById(id);
   }
 }
